style(cli): change lint code style
affects: @fect-ui/cli, @fect-ui/vue
