+++ packages/cli/src/commands/index.ts
@@ -3,6 +3,5 @@ export * from './compile'
 export * from './collect'
 export * from './dev'
 export * from './test'
-export * from './lint'
 export * from './build'
 export * from './volar'
